Replaces raw module for shell module
The built-in module used to gather facts is called `raw`, despite the type of command that is being performed.
There’s a difference between `raw`, `shell` and `command`.
I had to modify a few of the tasks to `shell` as Molecule is really strict on correct modules being used to do the right things.