Don't throw away a fingerprint value if it's None
Change current behavior to collect all facts, even when they are null.

The key change is the removal of condition on the function
`_add_fact_to_fingerprint`, allowing the desired behavior. This
required adaptations on `_compute_system_creation_time` and
`_merge_fingerprint` functions. It was also necessary to tweak
`SystemFingerprint` model and its serializer counterpart to accomodate
"null" values.

Also supress some rules for docstrings on migration files - instead of
documenting a file that's most of the time autogenerated, refer to
official django documentation.

Prerequisite for DISCOVERY-21 related subtsks.

Co-authored-by: Nicole Aragao <nicolearagao@gmail.com>