Integration test for OpenShift scans
Integration test based on the Network scan integration test. One important
distinction is that this new test is using django test utilities, not a
container. This was required in order to be able to patch OpenShiftApi
methods.

It'd be interesting to find a way to DRY up these tests since they have
a lot in common (at the moment it'd take a lot of effort given the
disntinct scope levels their fixtures use).

Part of DISCOVERY-187 implementation.

Signed-off-by: Bruno Ciconelle <bciconel@redhat.com>