feat(db-compat): Adapt psql StringAgg to SQLite
The next commit will use postgres StringAgg to optimize a query.
In order to keep our codebase compatible with SQLite, SQLiteStringAgg
was created.

It shouldn't be used directly, though. Instead use compat.StringAgg
will choose the appropriate aggregator class.

BTW, ArrayAgg would be more adequate for this task, but I couldn't
figure a way to reproduce it in SQLite.

Signed-off-by: Bruno Ciconelle <bciconel@redhat.com>