feat(ocp-api): dynamic client
Add a DynamicClient instance to OpenShiftApi.
This special client can call any type of k8s/ocp resource
given its api version and name.

To make testing easier, "private" attribute "_discoverer_cache_file" was
added.

In order to faccilitate the (re)creation of vcr cassettes the option
--refresh-cassetes was added. Tests marked with "vcr_primer" will be
the only ones executed with this option, and their respective cassettes
should be recreated when appropriate (conftest.py was moved to root folder
in order to allow adding this new option to pytest).

Co-authored-by: Nicole Aragão <nicolearagao@gmail.com>
Signed-off-by: Bruno Ciconelle <bciconel@redhat.com>