fix(test): OCP Tests are failing on MacOS
OCP Tests were failing on MacOS where Mocks were not being
honored and actual OpenShiftApi were attempted to test
OpenShift clusters and timing out.

The issue at hand is that on MacOS, multiprocess children
are spawned whereas the same are forked on Linux. With spawned
subprocesses, new Python interpreters are invoked which
lose the Mocked calls so we end up attempting to access the test
OpenShift clusters which do fail and eventually timeout.

The fix is to update the multiprocessing start method
to 'fork' for such cases so the Mocks are honored on MacOS.
At the end of the tests, the start methods are reverted
to the default ones (so back to spawn on MacOS).