fix(scans): Scans raise exceptions when running on MacOS.
Starting scans on MacOS cause an exception pertaining
to the Django app not properly initialized and unable
to use gettext for defining messages.

The multiprocessing package was updated in Python 3.8
on MacOS to use spawn instead of fork.

The issue is that for the scanner jobs, we create a Process
for each. With spawn, the Django environment is not passed
through and we would need to do a django.setup() in each
of the new subprocess.

Until multiprocessing is fixed in Python or we shift to something
like Celery which properly handles multiprocessing (via billiard)
the cleanest fix for MacOS is to switch the multiprocessing
start method to using fork early in the manage.py steps
so the change is less intrusive.