fix(sha256): Change how/when sha256 is calculated
The original function calculating SHA256 was doing it in a efficient way
(by avoiding loading all file content), but the overall process had 3
flaws that lead to the bug highlighted on previous commit:
- hash calculation is done before the files are actually created
- file was being created twice (one for hash calculation and then again
in "create_tar_buffer" function)
- in the end the whole file is loaded on memory anyway during tar file
generation, defeating the purpose of the efficiency on sha calculation.

Besides fixing the bug on sha generation, the changes introduced here to
"create_tar_buffer" future-proof it against this type of misuse.

Fix: DISCOVERY-265