refactor(CredentialSerializer): restore pre-refactor behavior
Restore pre-refactor validation for unexpected keys.

In order to preserve the original UX for CrededentialSerializer,
this weird validation is added: It will complain about extra fields, but
ONLY if they have a non-null value. A very unusual middle-ground.

I would rather go for a very strict validation, but that would break
the interaction between CLI and API.