fix: wrong key was being retrieved from output
Output from stdout_lines key was being retrieved instead of stdout key,
causing an implementation error.

Unit tests were not able to detect this error, on that account we decided to add this field to the integration test.
One of the required dependencies for the collection of installed_products is rct, therefore
it needs to be added to the scan-target dockerfile if we want to detect
a non-null value on the integration test.