fix: Allow git to run in /app
With the Exception catch removed while doing a git rev-parse for determining
the commit sha for the server version, git fails with the following
error while running in OpenShift:

  fatal: detected dubious ownership in repository at '/app'

This pertains to file/directory ownership vs the user running the git
command.  The fix to this is to tell git that /app is a safe
directory to run git in and proceed normally. The git config
adds /app as a safe directory to the user's .gitconfig file.