feature: create script to generate sudo cmds list
This commit introduces a new script that analyzes Ansible playbooks to collect all commands
requiring sudo permissions for execution. Additionally, it includes the following enhancements:

Integration of a step in GitHub Actions to ensure the generated list is up to date.
Addition of commands to the Makefile for generating the sudo command list and performing list comparison tests.