DISCOVERY-435: Authorize GitHub requests if token is available
On internal Jenkins, `make build-container` may fail because GitHub
rate-limits API requests, and a lot of internal traffic appears as
a single IP to GitHub.

To solve this, Jenkins will start providing API token when building
container. Obviously we don't want to require everyone to use a token,
so this is optional.

Token is provided through docker / podman `--secret` flag. That flag
must point to a file and that file must exist. So we provide empty one
here in commit, and will fill it on Jenkins.