Make Jenkins agent labels consistent with internally defined jobs
As we might get RHEL agents in the future (once qpc RPM is available),
it might be necessary to be more specific when selecting agent machine -
i.e. run something else on RHEL than on Fedora. This change ensures that
Fedora job is not mistakenly picked up by RHEL agent.