fix: Work around dnf grouplist returning early on unknown groups
dnf introduced backwards incompatible change - `dnf grouplist` will
return early if any argument is not a valid group name. This means we
can't ask for status of multiple groups at once and see which one is
installed - if any is not a valid group identifier, dnf will not even
check what is the status of others.

Relates to JIRA: DISCOVERY-456