refactor(Report): deprecate cached details report (sources field)
Deduplicate raw facts with the removal of sources field. Also add a
new model (InspectGroup) that has metadata once available on sources
and inexistent in the relational models.

As a replacement, add the cached_property sources to Report model.

Also, all scan related data is wiped out instead of being migrated
to the new model. The team made this choice because 1) the use case
for discovery is a customer that stands it up, perform scans and then
tears it down, and 2) fine tuning a proper data migration would
consume too much time for little value considering 1.

Relates to JIRA: DISCOVERY-218