refactor: split arm64 build to separate workflow
This is probably not ideal, but we want to run these builds in parallel (see
also the preceding commit that uses a matrix) and we need to track GitHub
action completion externally in discovery-ci to trigger test automation. We
suspect that changes to discovery-ci will be more straightforward if the
builds live in different workflow files.