refactor(report-merge)!: remove redundant views
Remove redundant views for getting status of a merge job (a merge job is still
a scan job, which can be checked in job viewset), and redundancy with report
upload view.

Also change the verb to "POST" as this view is creating something and returning 201 status code.

BREAKING CHANGE: deprecated verbs and replaced PUT with POST.

Relates to JIRA: DISCOVERY-452