fix: fail bigger when failing to access DeploymentsReport cached_fingerprints or cached_csv
After reflecting on the initial change introduced by 8729eb4d, we decided
that we would prefer to log errors and fail bigger if we fail to read
data from DeploymentsReport cached_fingerprints or cached_csv. This can
happen if the cached files were moved after they were written or if the
process trying to access the data does not have access to the files.
The latter could easily happen in a misconfigured container orchestration.

Following this change, if you perform a successful scan but then manually
destroy, move, or make unreadable its cached files, the next time you try
to download the report, you should encounter a 500 Internal Server Error.
Returning a 500 error is appropriate here because there is nothing the
client can do to fix the issue, and it requires some investigation and
unspecified fixes on the server side.

Relates to JIRA: DISCOVERY-703