fix: connect scan should fail if no hosts are reachable
This resolves a longstanding issue where the UI would present an apparent
success status for a network-type connection scan even if all of the hosts
in the source were unreachable. Previously, we would only fail if an
individual ansible run (hosts are split into groups, and run is invoked once
for each group) returned an edge-case final status like "timeout".

Relates to JIRA: DISCOVERY-823