test(network-scan): patch a flaky test
The sanity check for network scan fingerprints was failing
due to a flakiness with expected fingerprints. The full story is
"system_creation_date" fact is only present when one of its raw facts
has a non-null value. The test was failing when raw_facts_generator
had all their values set to null[1], which has a chance of ~12% of
ocurring.

Relates to JIRA: DISCOVERY-783

[1]: https://github.com/quipucords/quipucords/blob/6a935a47943366a2e8d0a0518f89d67acc009366/quipucords/tests/utils/raw_facts_generator.py#L124-L126