fix: declare and explicitly add poetry plugin poetry-plugin-export
This is necessary because the plugin is no longer included in Poetry 2.
Although plugins can be declared as dependencies and automatically installed
by inoking `poetry install`, we do not want to install all of our project's
dependencies just to lock and export requirements. There appears not to be
another way to install only the plugins via the declared dependencies in
pyproject.toml without installing *all* of the dependencies. So, we keep
this explicit step only for the "Check dependencies" GitHub CI workflow.