feat: add report ID to each JSON and CSV filename in downloaded tarball
This should help users who are manually processing files from multiple report directories.
That use case came to us as a specific feature request from an end user.

Upon expanding a report tarball download, the contents should now look like this example:

    report_id_3
    ├── aggregate-3.json
    ├── deployments-3.csv
    ├── deployments-3.json
    ├── details-3.csv
    ├── details-3.json
    ├── scan-job-3-ansible-stderr.txt
    ├── scan-job-3-ansible-stdout.txt
    └── SHA256SUM

In this example, the `-3` suffix is newly added to the aggregate, deployments, and details CSV and JSON filenames, and that value is the same as the report ID used in the parent directory's name and the downloaded tarball filename.

Relates to JIRA: DISCOVERY-788