feat: add report ID to ansible log filenames in report tarball
Upon expanding a report tarball download, the contents should now look like this example:

    report_id_18578
    ├── aggregate-18578.json
    ├── deployments-18578.csv
    ├── deployments-18578.json
    ├── details-18578.csv
    ├── details-18578.json
    ├── scan-job-18137-ansible-stderr-18578.txt
    ├── scan-job-18137-ansible-stdout-18578.txt
    └── SHA256SUM

The key change here from the previous behavior is the addition of the report ID suffix to the two `scan-job` files which is `-18578` in this example. However, please note that in this example, the `-18137-` values in the *middle* of the ansible logs filenames refers to the scan job's ID, not the report ID, and that value was already present in these filenames before this change.

Relates to JIRA: DISCOVERY-788