chore(Vagrantfile): drop generic RHEL 6 image
It seems there's something wrong with RHEL 6 image as it is not passing a verification (see the message below). Because this isn't
required for local testing (nor is RHEL 6 support official at this point),
let's simply ignore it.

An unexpected error occurred when executing the action on the
'rhel6' machine. Please report this as a bug:

EVP_DigestSignInit: invalid digest