fix: HostEntity.provider_id should return /etc/machine_id or another unique value to satisfy new HBI requirements
If a system fingerprint includes a value from `/etc/machine_id`, we will use that as `HostEntity.provider_id`. Else, we hash a set of other fingerprint facts with the expectation that they will be sufficiently unique to identify that system.

However, this change is not that simple! HBI also recently revised the list of attributes that HBI uses to determine uniqueness. We reproduced logic locally to define a list of "canonical facts". Before generating payload for HBI, we would remove any hosts that did not have at least one canonical fact. Because this change now *always* results in a not-empty `provider_id` value and `provider_id` is an attribute HBI now uses to determine uniqueness, this change also removes the previous logic around canonical facts and the filtering out of hosts that did not have at least one canonical fact.

Relates to JIRA: DISCOVERY-1063