chore: simplify logging configuration
There doesn't seem to be good reason for the very detailed logging
configuration that we had, which dates back to at least 2017. The result
is that as we add or rename modules, they might end up not having
logging configured - they will be handled by root logger, which does not
have nice formatter and by default prints only warning and above.

Switch logging configuration to more generic names. If we want to
silence some specific submodules, that should be done explicitly.